Moved timer printing code into display module
