+++ src/display.rs
@@ -2,8 +2,85 @@ use crate::args::DisplayCommand;
 use crate::error::{Error, Result};
 use crate::timer::TimerType;
 use chrono::{DateTime, Local, TimeZone};
+use crossterm::{
+    cursor, queue, style,
+    style::{Color, Stylize},
+    terminal::{Clear, ClearType},
+};
+use std::io::{stdout, Write};
 
-// TODO: Move the timer printing code into this module too, so our code for displaying to console
+// TODO: Put this in a submodule called timer?
+pub fn print_time_remaining(
+    time_remaining: u16,
+    total_time: u16,
+    timer_type: TimerType,
+) -> Result<()> {
+    let percentage: u64 = (100.0 - ((time_remaining as f64 / total_time as f64) * 100.0)) as u64;
+    let mut progress_bar: String = String::new();
+    let progress_amount = percentage / 10;
+    let space_amount = 10 - progress_amount;
+
+    if progress_amount > 0 {
+        for _i in 0..progress_amount {
+            progress_bar += "=";
+        }
+    }
+
+    if space_amount > 0 {
+        for _i in 0..space_amount {
+            progress_bar += " ";
+        }
+    }
+
+    let header = match timer_type {
+        TimerType::Work => String::from("Work Timer"),
+        TimerType::ShortBreak => String::from("Short Break Timer"),
+        TimerType::LongBreak => String::from("Long Break Timer"),
+    };
+
+    let minutes_component = time_remaining / 60;
+    let seconds_component = time_remaining % 60;
+
+    let mut stdout = stdout();
+    queue!(
+        stdout,
+        cursor::MoveTo(0, 0),
+        Clear(ClearType::FromCursorDown),
+        cursor::Hide,
+        style::Print(header),
+        cursor::MoveToNextLine(1),
+        style::Print(format_time(minutes_component, seconds_component)),
+        cursor::MoveToNextLine(1),
+        style::Print("["),
+        style::PrintStyledContent(progress_bar.with(Color::Green)),
+        style::Print(format!("] {}%", percentage))
+    )?;
+    stdout.flush()?;
+
+    Ok(())
+}
+
+pub fn print_timer_elapsed() -> Result<()> {
+    let mut stdout = stdout();
+    queue!(
+        stdout,
+        cursor::MoveToNextLine(1),
+        style::Print("Timer elapsed!"),
+        cursor::MoveToNextLine(1),
+        cursor::Show
+    )?;
+
+    Ok(())
+}
+
+fn format_time(minutes: u16, seconds: u16) -> String {
+    match seconds {
+        0..=10 => format!("{}:{:0>2} Remaining", minutes, seconds),
+        _ => format!("{}:{} Remaining", minutes, seconds),
+    }
+}
+
+// TODO: Put this in a submodule called session?
 pub fn print_sessions(
     sessions: Vec<(u64, u64)>,
     timer_type: TimerType,
